Wrapping in a lambda with application
Prevents the named let new definition from escaping the environment where it is used, like a correct let form should.
However the definition of make-lambda from the book seems difficult to use as you need to pass it a list of expressions such as '((+ x y)) for it to work.