Scanning out defines
The compiler did not implement let, so first I had to implement that by transforming the let with let->application (it becomes an immediately called lambda).
Then we are able to scan out the defines, which means to parse procedure bodies in compile-lambda-body and transform their defines in a big let which starts with the names being '*unassigned and then calls set! on each of them to assign the variable value (or lambda).